+++ Relay/Relay/Onboarding/RelayOnboardingViewController.swift
@@ -19,33 +19,6 @@ class RelayOnboardingViewController: UIViewController {
         return button
     }()
     
-    private let titleLabel: UILabel = {
-        let title = UILabel()
-        
-        title.text = "릴레이 설명"
-        title.font = .boldSystemFont(ofSize: 36)
-        
-        return title
-    }()
-    
-    private let contentLabel: UILabel = {
-        let content = UILabel()
-        
-        content.text = "어플을 설명하는 문장입니다\n어플을 설명하는 문장입니다\n어플을 설명하는 문장입니다"
-        content.numberOfLines = 3
-        content.font = .systemFont(ofSize: 17)
-        
-        return content
-    }()
-    
-    private let onboardingImageView: UIImageView = {
-        let imageView = UIImageView()
-        let myImage: UIImage = UIImage(named: "onboardingImage")!
-        imageView.image = myImage
-        
-        return imageView
-    }()
-    
     private let startButton: UIButton = {
         let button = UIButton()
         
@@ -79,9 +52,6 @@ class RelayOnboardingViewController: UIViewController {
     private func setupLayout(){
         [
             skipButton,
-            titleLabel,
-            contentLabel,
-            onboardingImageView,
             startButton,
             pageController
           
@@ -90,20 +60,6 @@ class RelayOnboardingViewController: UIViewController {
         skipButton.snp.makeConstraints {
             $0.top.equalToSuperview().inset(64.0)
             $0.trailing.equalToSuperview().inset(20.0)
-        }
-        titleLabel.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(160.0)
-            $0.leading.equalToSuperview().inset(50.0)
-        }
-        contentLabel.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom).offset(43.0)
-            $0.leading.equalToSuperview().offset(50.0)
-        }
-        onboardingImageView.snp.makeConstraints {
-            $0.top.equalTo(contentLabel.snp.bottom).offset(105.0)
-            $0.centerX.equalToSuperview()
-            $0.width.equalTo(116.0)
-            $0.height.equalTo(125.0)
         }
         startButton.snp.makeConstraints {
             $0.bottom.equalToSuperview().inset(75.0)
